+++ core/go/internal/sequencer/originator/originating.go
@@ -17,34 +17,15 @@ package originator
 
 import (
 	"context"
-	"fmt"
 
 	"slices"
 
-	"github.com/LFDT-Paladin/paladin/common/go/pkg/i18n"
 	"github.com/LFDT-Paladin/paladin/common/go/pkg/log"
 	"github.com/LFDT-Paladin/paladin/core/internal/components"
-	"github.com/LFDT-Paladin/paladin/core/internal/msgs"
 	"github.com/LFDT-Paladin/paladin/core/internal/sequencer/common"
 	"github.com/LFDT-Paladin/paladin/core/internal/sequencer/originator/transaction"
-	engineProto "github.com/LFDT-Paladin/paladin/core/pkg/proto/engine"
 	"github.com/google/uuid"
 )
-
-func validator_IsDelegationBlockHeightRejection(_ context.Context, _ *originator, event common.Event) (bool, error) {
-	return event.(*DelegationRequestRejectedEvent).RejectionReason == engineProto.RejectionReason_BLOCK_HEIGHT_TOLERANCE, nil
-}
-
-func validator_IsDelegationNotActiveCoordinatorRejection(_ context.Context, _ *originator, event common.Event) (bool, error) {
-	return event.(*DelegationRequestRejectedEvent).RejectionReason == engineProto.RejectionReason_NOT_CURRENT_DELEGATE, nil
-}
-
-func action_LogDelegationBlockHeightRejection(ctx context.Context, _ *originator, event common.Event) error {
-	e := event.(*DelegationRequestRejectedEvent)
-	log.L(ctx).Warnf("delegation rejected due to block height tolerance exceeded: originator block height=%d, coordinator block height=%d, coordinator tolerance=%d",
-		e.OriginatorBlockHeight, e.CoordinatorBlockHeight, e.BlockHeightTolerance)
-	return nil
-}
 
 func action_TransactionCreated(ctx context.Context, o *originator, event common.Event) error {
 	e := event.(*TransactionCreatedEvent)
@@ -102,110 +83,6 @@ func (o *originator) addToTransactions(
 	return nil
 }
 
-// sendDelegationRequest delegates transactions to the current active coordinator in the order they
-// were created on the originating node.
-//
-// When full is true, every resolved transaction in the contiguous prefix is (re)delegated. This is
-// used by the recovery paths (entry to Sending, dropped transactions, silence/failover, and
-// delegation-rejected redirect) where the coordinator may be missing state, so we bias toward
-// over-sending.
-//
-// When full is false only transactions whose state is State_Pending or State_Delegated are included,
-// i.e. the minimal set required to ensure FIFO ordering until first assembly
-func sendDelegationRequest(ctx context.Context, o *originator, full bool) error {
-	// Delegate the contiguous resolved prefix of transactions in the order they were created on the
-	// originating node. A still-resolving transaction blocks all later transactions so the coordinator
-	// receives them in creation order and never sees a transaction whose verifiers are unresolved.
-	inFlight := 0
-	transactionsToDelegate := make([]*components.PrivateTransaction, 0)
-	for _, txn := range o.transactionsOrdered {
-		// A transaction has advanced past verifier resolution, and so is eligible for delegation, only
-		// once it has left State_Initial and State_Resolving.
-		state := txn.GetCurrentState()
-		if state == transaction.State_Initial || state == transaction.State_Resolving {
-			break
-		}
-		inFlight++
-		// On the golden path skip any transaction the coordinator already knows about (it has been
-		// assembled at least once, so its ordering is locked in). We skip both the DelegatedEvent and
-		// the protobuf entry.
-		if !full && state != transaction.State_Pending && state != transaction.State_Delegated {
-			continue
-		}
-		transactionsToDelegate = append(transactionsToDelegate, txn.GetPrivateTransaction())
-		err := txn.HandleEvent(ctx, &transaction.DelegatedEvent{
-			BaseEvent: transaction.BaseEvent{
-				TransactionID: txn.GetID(),
-			},
-			Coordinator: o.currentActiveCoordinator,
-		})
-		if err != nil {
-			msg := fmt.Errorf("error handling delegated event for transaction %s: %v", txn.GetID(), err)
-			return i18n.NewError(ctx, msgs.MsgSequencerInternalError, msg)
-		}
-	}
-
-	if len(transactionsToDelegate) == 0 {
-		log.L(ctx).Debugf("no resolved transactions to delegate")
-		return nil
-	}
-
-	log.L(ctx).Debugf("sending delegation request for %d of %d in-flight transactions (full=%t)",
-		len(transactionsToDelegate), inFlight, full)
-
-	delegations := make([]*engineProto.PrivateTransactionDelegation, 0, len(transactionsToDelegate))
-	for _, tx := range transactionsToDelegate {
-		delegations = append(delegations, &engineProto.PrivateTransactionDelegation{
-			Id:          tx.ID.String(),
-			Domain:      tx.Domain,
-			Intent:      tx.Intent,
-			PreAssembly: tx.PreAssembly,
-		})
-	}
-	return o.transportWriter.SendDelegationRequest(ctx, o.currentActiveCoordinator, &engineProto.DelegationRequest{
-		DelegateNodeId:        o.currentActiveCoordinator,
-		OriginatorBlockHeight: int64(o.currentBlockHeight),
-		ContractAddress:       o.contractAddress.HexString(),
-		Transactions:          delegations,
-	})
-}
-
-// action_SignalDelegateNew raises the partial delegation dirty flag. The batching goroutine coalesces
-// this with any other signals in the current batch window and flushes a single partial delegation
-// (only transactions State_Pending / State_Delegated), preserving FIFO ordering until first assembly.
-func action_SignalDelegateNew(_ context.Context, o *originator, _ common.Event) error {
-	o.signalDelegate(false)
-	return nil
-}
-
-// action_SignalDelegateAll raises the full delegation dirty flag. The batching goroutine coalesces
-// this into a single full delegation (all unconfirmed transactions) on the next batch tick.
-func action_SignalDelegateAll(_ context.Context, o *originator, _ common.Event) error {
-	o.signalDelegate(true)
-	return nil
-}
-
-// action_FlushDelegation is the sole handler of Event_DelegateFlush, queued by the batching goroutine.
-// It refreshes the block height once per flush (rather than once per trigger) and sends the coalesced
-// delegation request.
-func action_FlushDelegation(ctx context.Context, o *originator, event common.Event) error {
-	e := event.(*DelegateFlushEvent)
-	o.refreshBlockHeight(ctx)
-	return sendDelegationRequest(ctx, o, e.Full)
-}
-
-// action_StartDelegationLoop starts the delegation batching goroutine on entry to State_Sending.
-func action_StartDelegationLoop(_ context.Context, o *originator, _ common.Event) error {
-	o.startDelegationLoop()
-	return nil
-}
-
-// action_StopDelegationLoop stops the delegation batching goroutine on exit from State_Sending.
-func action_StopDelegationLoop(_ context.Context, o *originator, _ common.Event) error {
-	o.stopDelegationLoop()
-	return nil
-}
-
 // action_RefreshBlockHeight queries the live block height and updates effectiveBlockHeight and the
 // priority list if the epoch has changed.
 func action_RefreshBlockHeight(ctx context.Context, o *originator, _ common.Event) error {
@@ -244,7 +121,12 @@ func action_FailoverToNextCoordinator(ctx context.Context, o *originator, _ comm
 		log.L(ctx).Debugf("originator failing over from %s to %s (failoverIndex now %d)",
 			prev, o.currentActiveCoordinator, o.failoverIndex)
 	}
-	o.signalDelegate(true)
+	// Notify the batching loop that a full delegation is required. Sending on a nil channel is never
+	// ready, so this is a safe no-op if the loop is not running.
+	select {
+	case o.notifyFullDelegation <- struct{}{}:
+	default:
+	}
 	return nil
 }
 
@@ -447,20 +329,6 @@ func action_UpdateActiveCoordinatorFromHeartbeat(_ context.Context, o *originato
 	e := event.(*common.HeartbeatReceivedEvent)
 	o.currentActiveCoordinator = e.FromNode
 	o.resetFailoverIndex()
-	return nil
-}
-
-// action_HandleDelegationRejected processes a rejection from a coordinator. If the rejection names
-// a coordinator that has higher priority than our current one, we redirect to it
-func action_HandleDelegationRejected(_ context.Context, o *originator, event common.Event) error {
-	e := event.(*DelegationRequestRejectedEvent)
-	if e.ActiveCoordinator == "" {
-		return nil
-	}
-	if common.IsHigherPriority(o.coordinatorPriorityList, e.ActiveCoordinator, o.currentActiveCoordinator) {
-		o.currentActiveCoordinator = e.ActiveCoordinator
-		o.resetFailoverIndex()
-	}
 	return nil
 }
